+++ R/predict_occ_taxon.R
@@ -17,8 +17,9 @@
 #'
 #' NOTE: to use this function you must first run `elements::startup()`
 #'
-#' @param taxon The taxon_code, see `elements::TaxonomicBackbone`.
-#' @param predictors A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, and bio17
+#' @param taxon The taxon_code, see `elements::ModelledTaxaCodes`.
+#' @param predictors A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, tmax_sm, tmin_wt, prec_wt, and prec_sm
+#' @param models description
 #' @param pa One of "Present", "Absent", or c("Present", "Absent").
 #' @param limit A string representing the niche width quantiles, one of "min_max", "q01_q99", "q05_q95", "q10_q90", "q25_q75". Which if set assigns a probability of 0 to the Present column and/or 1 to the Absent column to a set of predictors if one or more of those predictors are outside the stipulated quantile ranges. Optional.
 #' @param holdopt Hold one or more variables at their optimum values. NULL by default, else a vector of variable codes, e.g. c("SD", "GP").
@@ -34,7 +35,7 @@
 #' elements::predict_occ_taxon(taxon = "stellaria_graminea", predictors = elements::ExampleData1)
 #' elements::shutdown() 
 #' }
-predict_occ_taxon <- function(taxon, predictors, pa = "Present", limit = NULL, holdopt = NULL, dp = 3, append = "ids"){
+predict_occ_taxon <- function(taxon, predictors, models = elementsEnv$Models, pa = "Present", limit = NULL, holdopt = NULL, dp = 3, append = "ids"){
   
   # Check whether elements::startup() has been run and the Models filehashDB1 object is in the global environment
   if(isFALSE(exists(x = "Models", envir = elementsEnv))){
@@ -43,7 +44,7 @@ predict_occ_taxon <- function(taxon, predictors, pa = "Present", limit = NULL, h
   
   # Check whether all variables names are present in either 1) the predictors data frame, or 2) the holdopt argument
   if(!all(sort(unique(c(colnames(predictors[names(predictors) %in% elements::VariableNames]), holdopt))) == elements::VariableNames)){
-    stop("All model variables (L, M, N, R, S, SD, GP, bio05, bio06, bio16, bio17) must either be present in the predictors data frame or passed to holdopt.")
+    stop("All model variables (L, M, N, R, S, SD, GP, tmax_sm, tmin_wt, prec_wt, prec_sm) must either be present in the predictors data frame or passed to holdopt.")
   }
   
   # Check whether the limit argument is correct.
@@ -66,15 +67,15 @@ predict_occ_taxon <- function(taxon, predictors, pa = "Present", limit = NULL, h
   
   predictors_df <- predictors[elements::VariableNames]
   
-  model <- elementsEnv$Models[[taxon]]
+  model <- models[[taxon]]
   
-  predictions <- e1071:::predict.svm(object = model, predictors, probability = TRUE)
+  predictions <- e1071:::predict.svm(object = model, predictors_df, probability = TRUE)
   
   predictions_df <- round(as.data.frame(attr(predictions, "probabilities")[, pa, drop = FALSE]), digits = dp)
   
   if(!is.null(limit)){
     
-    predictors_check_lim <- elements::envelope_filter_taxon(taxon = taxon, predictors = predictors, limit = limit)
+    predictors_check_lim <- elements::envelope_filter_taxon(taxon = taxon, predictors = predictors_df, limit = limit)
     
     predictions_limited <- cbind(predictors_check_lim, predictions_df)
       
@@ -95,7 +96,7 @@ predict_occ_taxon <- function(taxon, predictors, pa = "Present", limit = NULL, h
   }
   
   if(append == "all"){
-    results_final <- cbind(predictors, results_final)
+    results_final <- cbind(predictors_df, ids_df, results_final)
   } else if(append == "predictors"){
     results_final <- cbind(predictors_df, results_final)
   } else if(append == "ids"){