+++ tests/testthat/test-env_filter.R
@@ -3,7 +3,7 @@ testthat::test_that("env_filter works with svm method and screen = TRUE", {
   elements::startup()
   
   test_predictors <- elements::ExampleScenarios
-  test_taxa <- elements::TaxonomicBackbone$taxon_code[sample(1:nrow(elements::TaxonomicBackbone), 20)]
+  test_taxa <- elements::ModelledTaxaCodes[sample(1:length(elements::ModelledTaxaCodes), 20)]
   test_method <- "svm"
   
   tictoc::tic()
@@ -12,7 +12,11 @@ testthat::test_that("env_filter works with svm method and screen = TRUE", {
   
   elements::shutdown()
   
-  testthat::expect_equal(colnames(actual), c("scenario", "timeslice", "scenario_code", "taxon_code", "Present"))
+  testthat::expect_true(setequal(colnames(actual), 
+                                 c(setdiff(colnames(test_predictors), elements::VariableNames), 
+                                   "taxon_code", "Present")
+                                 )
+                        )
   
 })
 
@@ -21,7 +25,7 @@ testthat::test_that("env_filter works with svm method and screen = FALSE", {
   elements::startup()
   
   test_predictors <- elements::ExampleScenarios
-  test_taxa <- elements::TaxonomicBackbone$taxon_code[sample(1:nrow(elements::TaxonomicBackbone), 20)]
+  test_taxa <- elements::ModelledTaxaCodes[sample(1:length(elements::ModelledTaxaCodes), 20)]
   test_method <- "svm"
   
   tictoc::tic()
@@ -30,42 +34,58 @@ testthat::test_that("env_filter works with svm method and screen = FALSE", {
   
   elements::shutdown()
   
-  testthat::expect_equal(colnames(actual), c("scenario", "timeslice", "scenario_code", "taxon_code", "Present"))
+  testthat::expect_true(setequal(colnames(actual), 
+                                 c(setdiff(colnames(test_predictors), elements::VariableNames), 
+                                   "taxon_code", "Present")
+  )
+  )
   
 })
 
 testthat::test_that("env_filter works with mean method", {
   
   test_predictors <- elements::ExampleScenarios
-  test_taxa <- elements::TaxonomicBackbone$taxon_code[sample(1:nrow(elements::TaxonomicBackbone), 20)]
+  test_taxa <- elements::ModelledTaxaCodes[sample(1:length(elements::ModelledTaxaCodes), 20)]
   test_method <- "mean"
   
   actual <- elements::env_filter(predictors = test_predictors, taxa = test_taxa, method = test_method)
   
-  testthat::expect_equal(colnames(actual), c("taxon_code", "scenario", "timeslice", "scenario_code", "distance"))
+  testthat::expect_true(setequal(colnames(actual), 
+                                 c(setdiff(colnames(test_predictors), elements::VariableNames), 
+                                   "taxon_code", "distance")
+                                 )
+                        )
   
 })
 
 testthat::test_that("env_filter works with median method", {
   
   test_predictors <- elements::ExampleScenarios
-  test_taxa <- elements::TaxonomicBackbone$taxon_code[sample(1:nrow(elements::TaxonomicBackbone), 20)]
+  test_taxa <- elements::ModelledTaxaCodes[sample(1:length(elements::ModelledTaxaCodes), 20)]
   test_method <- "median"
   
   actual <- elements::env_filter(predictors = test_predictors, taxa = test_taxa, method = test_method)
   
-  testthat::expect_equal(colnames(actual), c("taxon_code", "scenario", "timeslice", "scenario_code", "distance"))
+  testthat::expect_true(setequal(colnames(actual), 
+                                 c(setdiff(colnames(test_predictors), elements::VariableNames), 
+                                   "taxon_code", "distance")
+                                   )
+                        )
   
 })
 
 testthat::test_that("env_filter works with median method screen set to TRUE and limit set to min_max", {
   
   test_predictors <- elements::ExampleScenarios
-  test_taxa <-  elements::TaxonomicBackbone$taxon_code[sample(1:nrow(elements::TaxonomicBackbone), 20)]
+  test_taxa <-  elements::ModelledTaxaCodes[sample(1:length(elements::ModelledTaxaCodes), 20)]
   test_method <- "median"
   
   actual <- elements::env_filter(predictors = test_predictors, taxa = test_taxa, method = test_method, screen = TRUE, limit = "min_max")
   
-  testthat::expect_equal(colnames(actual), c("taxon_code", "scenario", "timeslice", "scenario_code", "distance"))
+  testthat::expect_true(setequal(colnames(actual), 
+                                 c(setdiff(colnames(test_predictors), elements::VariableNames), 
+                                   "taxon_code", "distance")
+                                 )
+                        )
   
 })