+++ tests/testthat/test-envelope_fillter_taxon.R
@@ -7,10 +7,10 @@ testthat::test_that("envelope_filter_taxon works", {
   
   actual <- elements::envelope_filter_taxon(taxon = test_taxon, predictors = test_predictors, vars = test_vars, limit = test_limit)
   
-  expected_colnames <- c("L", "M", "N", "R", "S", "SD", "GP", "bio05", "bio06", "bio16", "bio17", "within_limits")
+  expected_colnames <- c("L", "M", "N", "R", "S", "SD", "GP", "tmax_sm", "tmin_wt", "prec_wt", "prec_sm", "within_limits")
   expected_within_limits_values <- c(TRUE, FALSE)
   
-  testthat::expect_equal(colnames(actual), expected_colnames)
+  testthat::expect_true(setequal(colnames(actual), expected_colnames))
   testthat::expect_true(setequal(unique(actual$within_limits), expected_within_limits_values))
   
 })