+++ tests/testthat/test-predict_occ.R
@@ -12,19 +12,19 @@ testthat::test_that("predict_occ works", {
   # Generate predictions using a data frame containing taxa in the 'taxon_code' column.
   actual <- elements::predict_occ(taxa = NULL, predictors = test_predictors_2, pa = test_pa, limit = NULL, dp = 3, append = "all")
   
-  testthat::expect_equal(colnames(actual), c(colnames(test_predictors_2), test_pa))
+  testthat::expect_true(setequal(colnames(actual), c(colnames(test_predictors_2), test_pa)))
   testthat::expect_equal(nrow(actual), nrow(test_predictors_2))
   
   # Generate predictions for taxa specified in the 'taxa' argument using a data frame containing only predictor variables.
   actual <- elements::predict_occ(taxa= test_taxa, predictors = test_predictors_1, pa = test_pa, limit = NULL, dp = 3, append = "all")
   
-  testthat::expect_equal(colnames(actual), c(colnames(test_predictors_1), test_pa, "taxon_code"))
+  testthat::expect_true(setequal(colnames(actual), c(colnames(test_predictors_1), test_pa, "taxon_code")))
   testthat::expect_equal(nrow(actual), nrow(test_predictors_1) * 2)
   
   # Generate predictions for taxa specified in the 'taxa' argument using a data frame containing both predictor variables and a 'taxon_code' column (which is ignored and overwritten).
   actual <- elements::predict_occ(taxa = test_taxa, predictors = test_predictors_2, pa = test_pa, limit = NULL, dp = 3, append = "all")
   
-  testthat::expect_equal(colnames(actual), c(colnames(test_predictors_2), test_pa))
+  testthat::expect_true(setequal(colnames(actual), c(colnames(test_predictors_2), test_pa)))
   testthat::expect_equal(nrow(actual), nrow(test_predictors_2) * 2)