+++ tests/testthat/test-predict_occ_taxon.R
@@ -11,7 +11,7 @@ testthat::test_that("predict_occ_taxon works", {
   actual <- elements::predict_occ_taxon(taxon = test_taxon, predictors = test_predictors, 
                                         pa = test_pa, limit = NULL, holdopt = NULL, dp = 3, append = "all")
   
-  testthat::expect_equal(colnames(actual), c(colnames(test_predictors), test_pa))
+  testthat::expect_true(setequal(colnames(actual), c(colnames(test_predictors), test_pa)))
   testthat::expect_equal(nrow(actual), nrow(test_predictors))
   
   elements::shutdown()