+++ tests/testthat/test-startup.R
@@ -17,7 +17,7 @@ testthat:::test_that("startup works, models set to all", {
   
   models_names <- names(elementsEnv$Models)
   
-  testthat::expect_equal(sort(models_names), sort(elements::TaxonomicBackbone$taxon_code))
+  testthat::expect_true(setequal(models_names, elements::ModelledTaxaCodes))
   
   elements::shutdown()
   
@@ -29,7 +29,7 @@ testthat:::test_that("startup works, models set to test", {
   
   models_names <- names(elementsEnv$Models)
   
-  testthat::expect_equal(sort(models_names), sort(c("silene_flos-cuculi", "stellaria_graminea")))
+  testthat::expect_true(setequal(models_names, elements::ExamplePlot$taxon_code))
   
   elements::shutdown()